Crate lurk_macros

Lurk Macros

Derive macros for Lurk

This crate contains derive macros to manage trait dispatch in Lurk.

The Coproc macro adds dispatching Coprocessor and Cocircuit implementations to enums whose variants all atomically enclose types implementing Coprocessor.

Lurk macro

Although severely limited in the expressions it can represent, and still lacking quasiquoting, the lurk macro allows embedding Lurk code in Rust source. See tests for examples.

Macros

let_store

Binds a mutable Store to the variable s_, which is (somewhat unfortunately) hard-coded into the lurk! macro. Users should therefore avoid using s_ for other purposes, and will need to use s_ directly when manipulating the Store. The name s_ was chosen to walk the line between obscurity and brevity/clarity. Accidental collisions are undesirable, but so is an awkward or unwieldy name.

lurk

Derive Macros

Coproc